app/ai/requirements.txt
fastapi
uvicorn[standard]
huey
openai
python-dotenv
requests
tqdm 